requirements.txt
fastapi-users-db-sqlalchemy>=6.0.0
fastapi-users>=14.0.1
sqlalchemy>=2.0.38
fastapi<=0.115.8
pydantic-settings>=2.8.0
python-multipart>=0.0.9
email-validator>=2.2.0
pydantic>=2.10.6
starlette==0.45.2
pandas>=2.2.3
httpx==0.27.0
uvicorn>=0.29.0
fastapi-users[sqlalchemy]>=14.0.1
passlib[bcrypt]==1.7.4
aiosqlite
itsdangerous
jinja2
Pillow
markdown
openai == 1.69.0
python-dotenv
beautifulsoup4>=4.12.0
html2text>=2024.2.26
